Add files via upload
you will have to install the keyboard module and future module in the project. you can do that by entering the command in command promt "pip install [module]" than you have to find the module and place it in the ed2-unityproject folder. this is the way i tried to do.